fix: removed GitHubClient interface
